Remove `.zero()` and `.non_zero()` from `dsl::digit`
`.non_zero()` can be added back using `operator-`, if a user so desires.
It's no longer needed in the implementation of `dsl::digits`, so we're not supporting it.

`.zero()` is turned into a token `dsl::zero`.
It didn't depend on any of the existing bases, so why bother being more generic than necessary?