Workaround very specific GCC 7 potential miscompilation
If I use `dsl::lit_c<'\n'>` as the `newline` token, the context check for the second line of basic fails.
I can fix it by removing the `_choice` member typedef of `_whlc` (but then MSVC doesn't compile that one anymore).
This is really weird, as `_whlc` isn't instantiated in that file at all!
It's just sitting there, yet removing a member typedef fixes the test...?