Find longest instead of first match in alternative rule `/`
Originally, the alternative rule naively tried each token in order, backtracking if necessary.
Then I optimized it to use a trie for the literal parts;
this was technically a semantic change as `LEXY_LIT("a") / LEXY_LIT("abc")` suddenly matched `abc` (tries pick the longest match).
However, this optimization only applied if all the literals were at the beginning, i.e. `lit / lit / lit / other_token`,
and not for something like `lit / other_token / lit / lit` or even `other_token / lit / lit / lit`.

This commit started out with a fix for this optimization:
the type list is now (stable_)partitioned to move all the literals to the front to construct a trie and all other tokens to the back.
But this created intutive semantics with regards to ordering:
`LEXY_LIT("a") / while_(LEXY_LIT("b")) / LEXY_LIT("abc") / while_(dsl::ascii::alpha)` would match `abc` as the trie looks for the longest match,
but `bbc` as `bb` as the rest looks for the first.

So instead the logic is now made consistent to look for the longest match in either case.
As the alternative rule is mostly used for literals (maybe?), this should still be an optimization overall.
It is definitely more useful.