Make non-trivial `lexy::_detail::lazy_init` move-only
This worksaround a bug in GCC 7:
If a `lazy_init<T>` contains a `T` who has a `lazy_init<U>` as member,
where `U` is a non-trivial move-only type, the check for triviallity of
`T` instantiates `lazy_init<U>`'s copy constructor, which instantiates
`U`'s copy constructor, which doesn't work.

As we don't need to copy `lazy_init`, we might as well make it
move-only.