Change branch parsing of `dsl::identifier()`
It will now backtrack when trying to match a reserved identifier.
Previously, the branch was taken, an error raised, and immediately
recovered.

This new behavior is consistent with `dsl::symbol()` and also makes
semantic sense: when you want to parse an optional identifier,
a keyword shouldn't be considered to be an identifier.

The (common?) case where you're having a choice of multiple keywords and
then identifier as fallback, is unaffected by this change.