Don't eagerly find EOL in an `input_location::location`
This was problematic when combined with the `lexy_ext::shell`:
tracing prints the location where the final production ended, which is
at the next line.
As such, the eager EOL search in the constructor tries to find the end
of the line, which prompts for more input.

If we do it lazily when asking for `context()` or `newline()`, this is
avoided as trace don't need those.