Refactor parse context and handler
The handler is now split into an `event_handler` and a `value_callback`.
The former is essentially the old `marker` type, but with the `on` member functions,
the latter replaces the `list` event and takes the values from the `finish` event.

The parse context is now a dumb struct that contains the necessary data,
instead of a complex class with member functions.

This is both a significant simplification, and enables to easily change
the way values are produced during parsing.
The latter is necessary for the upcoming scan API.