Refactor `bp::try_parse()`
It now receives the control block only; not the whole context:
this is all the information necessary to decide on matching.
This simplifies the production branch parser.

This breaks `dsl::peek()` etc. which raise a bactracked event there.
As such, I've introduced a new `bp::cancel()` that must be called
instead of `bp::finish()` when `bp::try_parse()` returns false.
The event is now raised there.

This was also the motiviation behind the change: now `try_parse()` does
not generate any parse events, which allows an optimization in
`dsl::delimited()`.