Do not merge tokens in `lexy::token_production` anymore
The merging behavior is no longer desired with the new token kinds:
for example in JSON `1e-7` merges the `e` and `-` tokens into a single
literal.

`dsl::delimited()` now manually merges them, which was the main
situation where it was necessary.
This was only possible with the recent refactor.

As a drive-by change, error tokens are likewise not merged anymore,
which was also enabled by the refactor,
and the rules may now generate empty whitespace/error tokens, which the
actions filter out (this eliminates a branch if token events are
ignored).