Fix x86 build on Visual Studio (C++ 20 only)
This gets the Visual Studio build working for 32-bit builds for C++ 20 only.

Previously, the `_BitScanForward64()` function was being used for Visual Studio, which is only available in 64-bit builds.

This switches from `_BitScanForward64()` to `std::countr_zero()` for Visual Studio when C++ 20 or newer is used.

C++ 17 remains unsupported for 32-bit builds in Visual Studio.