cmake: do not exclusively set() in parent scope
under some external uses, cmake would error with:
> CMake Error (dev) at <...>/src/CMakeLists.txt:135 (target_sources):
>   uninitialized variable 'header_files'
> CMake Error (dev) at <...>/src/CMakeLists.txt:221 (target_sources):
>   uninitialized variable 'ext_header_files'

This is because the set() calls at the top of the file previously passed
PARENT_SCOPE, which does not only propagate these these vars to the
parent scope, but also results in them retaining an unset state in the
current scope. Quoting from the cmake docs for command/set:

> If the PARENT_SCOPE option is given the variable will be set in the
> scope above the current scope. [...] The previous state of the
> variable's value stays the same in the current scope (e.g., if it was
> undefined before, it is still undefined and if it had a value, it is
> still that value)

It's unclear to me why some build workflows encounter this and others do
not, as well as why the top-level CMakeLists.txt in lexy requires these
variables to be inserted into its scope (it does not seem to reference
them directly). So because I don't understand that, and in the interest
of not breaking any existing workflows while still fixing file-local
uses of these vars, add a second set() call which continues to pass
PARENT_SCOPE.

Signed-off-by: Nicholas Sielicki <opensource@nslick.com>